vita: fix overflow in lseek's return value
Fixes #129

+++ src/pkgi.hpp
@@ -139,7 +139,7 @@ void* pkgi_append(const char* path);
 
 void pkgi_close(void* f);
 
-int pkgi_seek(void* f, uint64_t offset);
+int64_t pkgi_seek(void* f, uint64_t offset);
 int pkgi_read(void* f, void* buffer, uint32_t size);
 int pkgi_write(void* f, const void* buffer, uint32_t size);
